@@ -1,15 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './App';
 import { ContextProvider } from './contexts/ContextProvider';
 
-ReactDOM.render(
+const root = createRoot(document.getElementById('root'));
+root.render(
   <React.StrictMode>
     <ContextProvider>
       <App />
     </ContextProvider>
   </React.StrictMode>,
-  document.getElementById('root'),
 );
